monitor/src/requirements.txt
aiohttp==3.7.4
bidict==0.21.2
click==8.0.1
crccheck==1.0
dnspython==1.16.0
docutils==0.17.1
eventlet==0.31.0
gevent==21.1.2
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
lockfile==0.12.2
MarkupSafe==2.0.1
mpmath==1.2.1
pymongo==3.11.4
pyserial==3.5
python-daemon==2.3.0
python-engineio==4.2.0
python-socketio==5.3.0
six==1.16.0
sympy==1.8
Werkzeug==2.0.1
zope.event==4.5.0
zope.interface==5.4.0
Flask==1.1.1
Flask-SocketIO==4.2.1
